Fix vulnerability dashboard filters
