+++ plugins/main/public/components/overview/vulnerabilities/dashboards/overview/dashboard-panels-kpis.ts
@@ -1,0 +1,415 @@
+import { DashboardPanelState } from '../../../../../../../../src/plugins/dashboard/public/application';
+import { EmbeddableInput } from '../../../../../../../../src/plugins/embeddable/public';
+import { VULNERABILITIES_INDEX_PATTERN_ID } from '../../common/constants';
+
+const getVisStateSeverityCritical = (indexPatternId: string) => {
+  return {
+    id: 'severity_critical_vulnerabilities',
+    title: 'Critical',
+    type: 'metric',
+    params: {
+      addTooltip: true,
+      addLegend: false,
+      type: 'metric',
+      metric: {
+        percentageMode: false,
+        useRanges: false,
+        colorSchema: 'Reds',
+        metricColorMode: 'Labels',
+        colorsRange: [
+          {
+            from: 0,
+            to: 0,
+          },
+          {
+            from: 0,
+            to: 0,
+          },
+        ],
+        labels: {
+          show: true,
+        },
+        invertColors: false,
+        style: {
+          bgFill: '#000',
+          bgColor: false,
+          labelColor: false,
+          subText: '',
+          fontSize: 50,
+        },
+      },
+    },
+    data: {
+      searchSource: {
+        query: {
+          language: 'kuery',
+          query: '',
+        },
+        filter: [],
+        index: indexPatternId,
+      },
+      references: [
+        {
+          name: 'kibanaSavedObjectMeta.searchSourceJSON.index',
+          type: 'index-pattern',
+          id: indexPatternId,
+        },
+      ],
+      aggs: [
+        {
+          id: '1',
+          enabled: true,
+          type: 'count',
+          params: {
+            customLabel: ' ',
+          },
+          schema: 'metric',
+        },
+        {
+          id: '2',
+          enabled: true,
+          type: 'filters',
+          params: {
+            filters: [
+              {
+                input: {
+                  query: 'vulnerability.severity:"critical"',
+                  language: 'kuery',
+                },
+                label: '- Critical Severity Alerts',
+              },
+            ],
+          },
+          schema: 'group',
+        },
+      ],
+    },
+  };
+};
+
+const getVisStateSeverityHigh = (indexPatternId: string) => {
+  return {
+    id: 'severity_high_vulnerabilities',
+    title: 'High',
+    type: 'metric',
+    params: {
+      addTooltip: true,
+      addLegend: false,
+      type: 'metric',
+      metric: {
+        percentageMode: false,
+        useRanges: false,
+        colorSchema: 'Blues',
+        metricColorMode: 'Labels',
+        colorsRange: [
+          {
+            from: 0,
+            to: 0,
+          },
+          {
+            from: 0,
+            to: 0,
+          },
+        ],
+        labels: {
+          show: true,
+        },
+        invertColors: false,
+        style: {
+          bgFill: '#000',
+          bgColor: false,
+          labelColor: false,
+          subText: '',
+          fontSize: 50,
+        },
+      },
+    },
+    uiState: {
+      vis: {
+        colors: {
+          'High Severity Alerts - Count': '#38D1BA',
+        },
+      },
+    },
+    data: {
+      searchSource: {
+        query: {
+          language: 'kuery',
+          query: '',
+        },
+        filter: [],
+        index: indexPatternId,
+      },
+      references: [
+        {
+          name: 'kibanaSavedObjectMeta.searchSourceJSON.index',
+          type: 'index-pattern',
+          id: indexPatternId,
+        },
+      ],
+      aggs: [
+        {
+          id: '1',
+          enabled: true,
+          type: 'count',
+          params: {
+            customLabel: ' ',
+          },
+          schema: 'metric',
+        },
+        {
+          id: '2',
+          enabled: true,
+          type: 'filters',
+          params: {
+            filters: [
+              {
+                input: {
+                  query: 'vulnerability.severity:"high"',
+                  language: 'kuery',
+                },
+                label: '- High Severity Alerts',
+              },
+            ],
+          },
+          schema: 'group',
+        },
+      ],
+    },
+  };
+};
+
+const getVisStateSeverityMedium = (indexPatternId: string) => {
+  return {
+    id: 'severity_medium_vulnerabilities',
+    title: 'Medium',
+    type: 'metric',
+    params: {
+      addTooltip: true,
+      addLegend: false,
+      type: 'metric',
+      metric: {
+        percentageMode: false,
+        useRanges: false,
+        colorSchema: 'Yellow to Red',
+        metricColorMode: 'Labels',
+        colorsRange: [
+          {
+            from: 0,
+            to: 0,
+          },
+          {
+            from: 0,
+            to: 0,
+          },
+        ],
+        labels: {
+          show: true,
+        },
+        invertColors: true,
+        style: {
+          bgFill: '#000',
+          bgColor: false,
+          labelColor: false,
+          subText: '',
+          fontSize: 50,
+        },
+      },
+    },
+    data: {
+      searchSource: {
+        query: {
+          language: 'kuery',
+          query: '',
+        },
+        filter: [],
+        index: indexPatternId,
+      },
+      references: [
+        {
+          name: 'kibanaSavedObjectMeta.searchSourceJSON.index',
+          type: 'index-pattern',
+          id: indexPatternId,
+        },
+      ],
+      aggs: [
+        {
+          id: '1',
+          enabled: true,
+          type: 'count',
+          params: {
+            customLabel: ' ',
+          },
+          schema: 'metric',
+        },
+        {
+          id: '2',
+          enabled: true,
+          type: 'filters',
+          params: {
+            filters: [
+              {
+                input: {
+                  query: 'vulnerability.severity:"medium"',
+                  language: 'kuery',
+                },
+                label: '- Medium Severity Alerts',
+              },
+            ],
+          },
+          schema: 'group',
+        },
+      ],
+    },
+  };
+};
+
+const getVisStateSeverityLow = (indexPatternId: string) => {
+  return {
+    id: 'severity_low_vulnerabilities',
+    title: 'Low',
+    type: 'metric',
+    params: {
+      addTooltip: true,
+      addLegend: false,
+      type: 'metric',
+      metric: {
+        percentageMode: false,
+        useRanges: false,
+        colorSchema: 'Greens',
+        metricColorMode: 'Labels',
+        colorsRange: [
+          {
+            from: 0,
+            to: 0,
+          },
+          {
+            from: 0,
+            to: 0,
+          },
+        ],
+        labels: {
+          show: true,
+        },
+        invertColors: false,
+        style: {
+          bgFill: '#000',
+          bgColor: false,
+          labelColor: false,
+          subText: '',
+          fontSize: 50,
+        },
+      },
+    },
+    data: {
+      searchSource: {
+        query: {
+          language: 'kuery',
+          query: '',
+        },
+        filter: [],
+        index: indexPatternId,
+      },
+      references: [
+        {
+          name: 'kibanaSavedObjectMeta.searchSourceJSON.index',
+          type: 'index-pattern',
+          id: indexPatternId,
+        },
+      ],
+      aggs: [
+        {
+          id: '1',
+          enabled: true,
+          type: 'count',
+          params: {
+            customLabel: ' ',
+          },
+          schema: 'metric',
+        },
+        {
+          id: '2',
+          enabled: true,
+          type: 'filters',
+          params: {
+            filters: [
+              {
+                input: {
+                  query: 'vulnerability.severity:"low"',
+                  language: 'kuery',
+                },
+                label: '- Low Severity Alerts',
+              },
+            ],
+          },
+          schema: 'group',
+        },
+      ],
+    },
+  };
+};
+
+export const getKPIsPanel = (): {
+  [panelId: string]: DashboardPanelState<
+    EmbeddableInput & { [k: string]: unknown }
+  >;
+} => {
+  return {
+    '1': {
+      gridData: {
+        w: 12,
+        h: 6,
+        x: 0,
+        y: 0,
+        i: '1',
+      },
+      type: 'visualization',
+      explicitInput: {
+        id: '1',
+        savedVis: getVisStateSeverityCritical(VULNERABILITIES_INDEX_PATTERN_ID),
+      },
+    },
+    '2': {
+      gridData: {
+        w: 12,
+        h: 6,
+        x: 12,
+        y: 0,
+        i: '2',
+      },
+      type: 'visualization',
+      explicitInput: {
+        id: '2',
+        savedVis: getVisStateSeverityHigh(VULNERABILITIES_INDEX_PATTERN_ID),
+      },
+    },
+    '3': {
+      gridData: {
+        w: 12,
+        h: 6,
+        x: 24,
+        y: 0,
+        i: '3',
+      },
+      type: 'visualization',
+      explicitInput: {
+        id: '3',
+        savedVis: getVisStateSeverityMedium(VULNERABILITIES_INDEX_PATTERN_ID),
+      },
+    },
+    '4': {
+      gridData: {
+        w: 12,
+        h: 6,
+        x: 36,
+        y: 0,
+        i: '4',
+      },
+      type: 'visualization',
+      explicitInput: {
+        id: '4',
+        savedVis: getVisStateSeverityLow(VULNERABILITIES_INDEX_PATTERN_ID),
+      },
+    },
+  };
+};
