+++ plugins/main/public/components/overview/vulnerabilities/dashboards/overview/vulnerability-detector-filters.scss
